pyproject.toml
[project]
name = "voice-subtitle-generator"
version = "0.1.0"
description = "Japanese audio to Korean subtitle generator with speaker diarization"
requires-python = ">=3.11"
dependencies = [
    # STT (Speech-to-Text)
    "faster-whisper>=1.0.0",

    # Speaker Diarization
    "pyannote-audio>=3.1",

    # LLM Translation
    "llama-cpp-python>=0.2.0",

    # Audio Processing
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    "torchaudio>=2.0.0",

    # Subtitle Generation
    "pysubs2>=1.6.0",

    # GUI
    "dearpygui>=1.11",

    # Configuration
    "pyyaml>=6.0",

    # Utilities
    "tqdm>=4.65.0",
    "scipy>=1.10.0",  # For clustering in speechbrain diarization
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "ruff>=0.1.0",
]
build = [
    "pyinstaller>=6.0",
]
speechbrain = [
    "speechbrain>=1.0.0",  # Alternative diarization backend (no HF token required)
]

# [project.scripts]
# vsg = "main:main"  # Run with: python main.py

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "ui"]

[tool.hatch.build.targets.sdist]
include = ["src", "ui", "config", "main.py"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
